ui: add conference contribution link to detail page
* ref: cern-sis/issues-inspire#443
